pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["abin <abin@vizru.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pymupdf = "^1.24.11"
langchain-openai = "^0.2.2"
python-dotenv = "^1.0.1"
openai = "^1.51.2"
langchain-text-splitters = "^0.3.0"
fastapi = "^0.115.2"
uvicorn = "^0.31.1"
python-multipart = "^0.0.12"
wasabi = "^1.1.3"
weaviate-client = "^4.8.1"
pymongo = "^4.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
